+++ src/bb-components-build.ts
@@ -141,7 +141,9 @@ const readComponents: () => Promise<Component[]> = async (): Promise<
   return Promise.all(components);
 };
 
-const readtsPrefabs: () => Promise<Prefab[]> = async (): Promise<Prefab[]> => {
+const readtsPrefabs: (isPartial?: boolean) => Promise<Prefab[]> = async (
+  isPartial = false,
+): Promise<Prefab[]> => {
   const absoluteRootDir = path.resolve(process.cwd(), rootDir);
   const srcDir = `${absoluteRootDir}/src/prefabs`;
   const outDir = `${absoluteRootDir}/tmp/${Math.floor(
@@ -154,13 +156,23 @@ const readtsPrefabs: () => Promise<Prefab[]> = async (): Promise<Prefab[]> => {
     throw new Error(chalk.red('\nPrefabs folder not found\n'));
   }
 
-  const prefabTsFiles: string[] = await readFilesByType(srcDir, 'ts');
-  const prefabTsxFiles: string[] = await readFilesByType(srcDir, 'tsx');
+  let prefabTsFiles: string[] = [];
+  let prefabTsxFiles: string[] = [];
+  let partialTsxFiles: string[] = [];
+
+  if (isPartial) {
+    partialTsxFiles = await readFilesByType(`${srcDir}/partials`, 'tsx');
+  } else {
+    prefabTsFiles = await readFilesByType(srcDir, 'ts');
+    prefabTsxFiles = await readFilesByType(srcDir, 'tsx');
+  }
 
   const prefabFiles = [...prefabTsFiles, ...prefabTsxFiles];
 
+  const files = isPartial ? partialTsxFiles : prefabFiles;
+  const basePath = isPartial ? `${srcDir}/partials/` : `${srcDir}/`;
   const prefabProgram = ts.createProgram(
-    prefabFiles.map((file) => `${srcDir}/${file}`),
+    files.map((file) => `${basePath}/${file}`),
     {
       allowSyntheticDefaultImports: false,
       esModuleInterop: true,
@@ -207,7 +219,11 @@ const readtsPrefabs: () => Promise<Prefab[]> = async (): Promise<Prefab[]> => {
   }
 
   const prefabs: Array<Promise<Prefab>> = (results.emittedFiles || [])
-    .filter((filename) => /prefabs\/\w+\.js$/.test(filename))
+    .filter((filename) =>
+      isPartial
+        ? /prefabs\/partials\/\w+\.js$/.test(filename)
+        : /prefabs\/\w+\.js$/.test(filename),
+    )
     .map((filename) => {
       return new Promise((resolve) => {
         import(filename)
@@ -342,6 +358,7 @@ void (async (): Promise<void> => {
       jsPrefabs,
       components,
       interactions,
+      tsxPartialPrefabs,
       partialprefabs,
     ] = await Promise.all([
       readStyles(rootDir),
@@ -349,12 +366,17 @@ void (async (): Promise<void> => {
       readPrefabs(),
       readComponents(),
       runtimeVersion === 'v2' ? Promise.resolve([]) : readInteractions(),
+      readtsPrefabs(true),
       readPartialPrefabs(),
     ]);
 
     const validStyleTypes = styles.map(({ type }) => type);
     const prefabs = jsPrefabs
       .concat(tsxPrefabs)
+      .filter((prefab): prefab is Prefab => !!prefab);
+
+    const allPartialPrefabs = partialprefabs
+      .concat(tsxPartialPrefabs)
       .filter((prefab): prefab is Prefab => !!prefab);
 
     const stylesGroupedByTypeAndName = styles.reduce<GroupedStyles>(
@@ -389,7 +411,7 @@ void (async (): Promise<void> => {
       validateComponents(components, validStyleTypes),
       validatePrefabs(prefabs, stylesGroupedByTypeAndName, componentStyleMap),
       validatePrefabs(
-        partialprefabs,
+        allPartialPrefabs,
         stylesGroupedByTypeAndName,
         componentStyleMap,
         'partial',
@@ -445,7 +467,7 @@ void (async (): Promise<void> => {
     const buildStyles = styles.map(buildStyle);
 
     const buildPrefabs = prefabs.map(buildPrefab);
-    const buildPartialprefabs = partialprefabs.map(buildPrefab);
+    const buildPartialprefabs = allPartialPrefabs.map(buildPrefab);
 
     await mkdir(distDir, { recursive: true });