+++ src/bb-components-publish.ts
@@ -4,25 +4,8 @@
 import program, { CommanderStatic } from 'commander';
 import chalk from 'chalk';
 import { readJSON, pathExists } from 'fs-extra';
-
-import uploadBlob, {
-  BlockBlobUploadResponseExtended,
-} from './utils/uploadBlob';
 import { checkUpdateAvailableCLI } from './utils/checkUpdateAvailable';
-
-/* setup */
-
-const { AZURE_BLOB_ACCOUNT, AZURE_BLOB_ACCOUNT_KEY } = process.env;
-
-if (typeof AZURE_BLOB_ACCOUNT !== 'string') {
-  throw new Error(chalk.red('\n$AZURE_BLOB_ACCOUNT is required\n'));
-}
-
-if (typeof AZURE_BLOB_ACCOUNT_KEY !== 'string') {
-  throw new Error(chalk.red('\n$AZURE_BLOB_ACCOUNT_KEY is required\n'));
-}
-
-/* process arguments */
+import { publish } from './functions/bb-components-functions';
 
 program
   .usage('[options] [path]')
@@ -36,8 +19,6 @@ const distDir: string = args.length === 0 ? 'dist' : `${args[0]}/dist`;
 if (!name || typeof name !== 'string' || !name.length) {
   throw new Error(chalk.red('\n-b or --bucket [name] is required\n'));
 }
-
-/* execute command */
 
 const read = async (fileName: string): Promise<void> => {
   try {
@@ -56,42 +37,6 @@ const read = async (fileName: string): Promise<void> => {
   }
 };
 
-const upload = async (
-  objects: unknown,
-  fileName: string,
-): Promise<BlockBlobUploadResponseExtended> => {
-  try {
-    return await uploadBlob(name, fileName, JSON.stringify(objects));
-  } catch (error) {
-    const defaultMessage =
-      'There was an error trying to publish your component set';
-    const { body, message } = error;
-
-    if (!body) {
-      throw new Error(chalk.red([defaultMessage, message].join('\n')));
-    }
-
-    const { code, message: bodyMessage } = body;
-
-    const extraMessage =
-      code === 'AuthenticationFailed'
-        ? 'Make sure your azure blob account and key are correct'
-        : bodyMessage;
-
-    throw new Error(chalk.red([defaultMessage, extraMessage].join('\n')));
-  }
-};
-
-const publish = async (
-  fileName: string,
-): Promise<BlockBlobUploadResponseExtended> => {
-  console.log(`Publishing ${fileName}.`);
-
-  const objects = await read(fileName);
-
-  return upload(objects, fileName);
-};
-
 // eslint-disable-next-line no-void
 void (async (): Promise<void> => {
   await checkUpdateAvailableCLI();
@@ -105,7 +50,9 @@ void (async (): Promise<void> => {
   if (existingPartialPath) {
     files.push('partials.json');
   }
-  const [{ url }] = await Promise.all(files.map(publish));
+  const [{ url }] = await Promise.all(
+    files.map((fileName) => publish(fileName, name, read)),
+  );
 
   console.log(
     chalk.green(