+++ src/validations/prefab/linkedPartial.ts
@@ -15,6 +15,7 @@ export const linkedPartialConfigurationSchema = Joi.object({
     comparator: Joi.string().valid(...COMPARATORS),
     value: Joi.any(),
   }),
+  showOnDrop: Joi.boolean(),
 });
 
 export const linkedPartialSchema = Joi.object({
@@ -23,4 +24,9 @@ export const linkedPartialSchema = Joi.object({
   type: Joi.string().valid('LINKED_PARTIAL').required(),
   value: linkedPartialValueSchema,
   configuration: linkedPartialConfigurationSchema,
+  optionRef: Joi.object({
+    id: Joi.string(),
+    sourceId: Joi.string(),
+    inherit: ['label', 'name', 'variableProperty'],
+  }),
 });